+++ package.json
@@ -20,15 +20,11 @@
     "@navikt/ds-tailwind": "^7.22.0",
     "date-fns": "^4.1.0",
     "dotenv": "^16.5.0",
-    "lodash": "^4.17.21",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-helmet": "^6.1.0",
     "react-markdown": "^9.1.0",
-    "react-responsive": "^10.0.1",
     "react-router": "^7.6.2",
-    "react-toggle-dark-mode": "^1.1.1",
-    "react-use-websocket": "^4.13.0",
     "rehype-raw": "^7.0.0",
     "rehype-stringify": "^10.0.1",
     "remark": "^15.0.1",
@@ -41,8 +37,6 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
-    "@types/express": "^5.0.2",
-    "@types/lodash": "^4.17.17",
     "@types/react": "^19.1.6",
     "@types/react-dom": "^19.1.5",
     "@types/react-helmet": "^6.1.11",
@@ -51,11 +45,7 @@
     "eslint": "^9.28.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
-    "express": "^4.21.2",
     "globals": "^16.2.0",
-    "http-proxy-middleware": "^3.0.5",
-    "node-ts-cache": "^4.4.0",
-    "node-ts-cache-storage-memory": "^4.4.0",
     "postcss": "^8.5.4",
     "prettier": "^3.5.3",
     "prettier-plugin-organize-imports": "^4.1.0",
